Fix client

+++ authentication/authentication/src/main/java/com/example/authentication/service/implement/MobileServiceImpl.java
@@ -45,7 +45,7 @@ public class MobileServiceImpl implements MobileService {
             put("mobileQuantity", mobileEntity.getMobileQuantity());
             put("mobilePrice", mobileEntity.getMobilePrice());
             put("mobileDescription", mobileEntity.getMobileDescription());
-            put("imageUrl", mobileEntity.getImageUrl());
+            put("imageUrl", mobileEntity.getImageUrl() != null ? mobileEntity.getImageUrl() : "");
             put("rating", mobileEntity.getRate());
         }};
     }
@@ -59,10 +59,12 @@ public class MobileServiceImpl implements MobileService {
             RatingEntity ratingEntity = new RatingEntity();
             ratingRepository.save(ratingEntity);
             mobile.setRate(ratingEntity);
-            if (!mobile.getImageUrl().isEmpty()) {
-                // Save to S3 Bucket
-                URL objectURL = s3Utils.getS3URL(mobile.getImageUrl());
-                mobile.setImageUrl(objectURL.toString());
+            if (mobile.getImageUrl() != null) {
+                if (!mobile.getImageUrl().isEmpty()) {
+                    // Save to S3 Bucket
+                    URL objectURL = s3Utils.getS3URL(mobile.getImageUrl());
+                    mobile.setImageUrl(objectURL.toString());
+                }
             }
 
             BeanUtils.copyProperties(mobile, mobileEntity);
@@ -149,9 +151,11 @@ public class MobileServiceImpl implements MobileService {
             mobileEntity.setMobilePrice(mobile.getMobilePrice());
             mobileEntity.setMobileDescription(mobile.getMobileDescription());
             mobileEntity.setUpdateAt(LocalDateTime.now());
-            if (!mobile.getImageUrl().isEmpty()) {
-                URL objectURL = s3Utils.getS3URL(mobile.getImageUrl());
-                mobileEntity.setImageUrl(objectURL.toString());
+            if (mobile.getImageUrl() != null) {
+                if (!mobile.getImageUrl().isEmpty()) {
+                    URL objectURL = s3Utils.getS3URL(mobile.getImageUrl());
+                    mobileEntity.setImageUrl(objectURL.toString());
+                }
             }
             mobileRepository.save(mobileEntity);
             BeanUtils.copyProperties(mobileEntity, mobile);
@@ -165,14 +169,18 @@ public class MobileServiceImpl implements MobileService {
     public Boolean deleteMobile(Long mobileId) throws Exception {
         try {
             if (mobileRepository.findById(mobileId).isPresent()) {
-                // Delete image in S3 Bucket
-                String fileURI = mobileRepository.findById(mobileId).get().getImageUrl();
-                String [] fileURISplitted = fileURI.split("/");
-                log.info("fileURISplitted: {}", (Object) fileURISplitted);
-                String fileName = fileURISplitted[fileURISplitted.length-1];
-                log.info("FileName: {}", fileName);
-                s3Client.deleteObject(bucketName, fileName);
-                log.info("FileName: {} removed", fileName);
+                if (mobileRepository.findById(mobileId).get().getImageUrl() != null) {
+                    if (!mobileRepository.findById(mobileId).get().getImageUrl().isEmpty()) {
+                        // Delete image in S3 Bucket
+                        String fileURI = mobileRepository.findById(mobileId).get().getImageUrl();
+                        String [] fileURISplitted = fileURI.split("/");
+                        log.info("fileURISplitted: {}", (Object) fileURISplitted);
+                        String fileName = fileURISplitted[fileURISplitted.length-1];
+                        log.info("FileName: {}", fileName);
+                        s3Client.deleteObject(bucketName, fileName);
+                        log.info("FileName: {} removed", fileName);
+                    }
+                }
 
                 mobileRepository.delete(mobileRepository.findById(mobileId).get());
                 return true;
